move test into SurveyApplicationTests
